separated workflow tab to its folder
